more changes

+++ Plutus/src/Assets/SoundLoader.h
@@ -4,9 +4,6 @@
 #include <fstream>
 #include <vector>
 
-#include <Utils/types.h>
-
-
 namespace Plutus {
 
 	struct SubFormat {
@@ -18,30 +15,30 @@ namespace Plutus {
 
 	//Copied from microsoft WAVEFORMATEX
 	struct WaveFormatEx {
-		u16        wFormatTag;         /* format type */
-		u16        nChannels;          /* number of channels (i.e. mono, stereo...) */
-		ulong       nSamplesPerSec;     /* sample rate */
-		ulong       nAvgBytesPerSec;    /* for buffer estimation */
-		u16        nBlockAlign;        /* block size of data */
-		u16        wBitsPerSample;     /* number of bits per sample of mono data */
-		u16        cbSize;             /* the count in bytes of the size of extra information (after cbSize) */
+		uint16_t        wFormatTag;         /* format type */
+		uint16_t        nChannels;          /* number of channels (i.e. mono, stereo...) */
+		uint32_t       nSamplesPerSec;     /* sample rate */
+		uint32_t       nAvgBytesPerSec;    /* for buffer estimation */
+		uint16_t        nBlockAlign;        /* block size of data */
+		uint16_t        wBitsPerSample;     /* number of bits per sample of mono data */
+		uint16_t        cbSize;             /* the count in bytes of the size of extra information (after cbSize) */
 		union {
-			u16 wValidBitsPerSample;       /* bits of precision  */
-			u16 wSamplesPerBlock;          /* valid if wBitsPerSample==0 */
-			u16 wReserved;                 /* If neither applies, set to zero. */
+			uint16_t wValidBitsPerSample;       /* bits of precision  */
+			uint16_t wSamplesPerBlock;          /* valid if wBitsPerSample==0 */
+			uint16_t wReserved;                 /* If neither applies, set to zero. */
 		} Samples;
-		ulong           dwChannelMask;      /* which channels are */
+		uint32_t           dwChannelMask;      /* which channels are */
 		SubFormat subformat;
 	};
 
 	class SoundLoader
 	{
 	public:
-		static bool loadSoundFile(const std::string& filePath, void* wfx, std::vector<u8>* buffer);
+		static bool loadSoundFile(const std::string& filePath, void* wfx, std::vector<uint8_t>* buffer);
 
 	private:
 		static bool readChunkToBuffer(std::string file, void* buffer, unsigned long buffSize, unsigned long offset);
 		static bool readWave(std::string filePath, void* wfx, unsigned long& buffSize, unsigned long& offset);
-		static bool readOgg(std::string filePath, void* wfx, std::vector<u8>* buffer);
+		static bool readOgg(std::string filePath, void* wfx, std::vector<uint8_t>* buffer);
 	};
 }
